requirements.txt
deepface
opencv-python
numpy
mediapipe
cryptography
PyYAML
tf-keras
# tensorflow==2.13.0
pytest